requirements.txt
blinker==1.9.0
Brotli==1.1.0
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cssselect2==0.8.0
Flask==3.1.1
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.58.4
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.2.1
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
python-dotenv==1.1.1
qrcode==8.2
reportlab==4.4.2
SQLAlchemy==2.0.41
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.14.0
weasyprint==65.1
webencodings==0.5.1
Werkzeug==3.1.3
WTForms==3.2.1
zopfli==0.2.3.post1
